# internal/server/assets/openapi.yaml
openapi: 3.1.0
info:
  title: GoDrop
  summary: Upload a file, get a hard-to-guess public URL.
  description: |
    A tiny self-hosted file host. Uploads and deletes require an API token;
    downloads are public. There is no listing endpoint: identifiers carry 128
    bits of entropy and are the only way to reach a file.
  version: "1.2.6"
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /
    description: This instance
security:
  - bearerAuth: []
paths:
  /upload:
    post:
      operationId: uploadFiles
      summary: Upload one or more files
      description: |
        Send `multipart/form-data` with one or more parts named `file`. The
        upload is atomic: if any part is rejected, previously written parts of
        the same request are deleted.
      parameters:
        - name: X-Expires-In
          in: header
          required: false
          description: >
            Delete the file after this long: 30m, 12h, 7d, 30d. Capped at the
            instance's retention period when it has one. ?expires= does the
            same thing for clients that cannot set a header.
          schema: { type: string, example: 7d }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
              required: [file]
      responses:
        "201":
          description: Files stored
          headers:
            Location:
              description: URL of the first file stored by this request.
              schema: { type: string, format: uri }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "415": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
        "507": { $ref: "#/components/responses/Error" }
  /upload/{name}:
    put:
      operationId: uploadRaw
      summary: Upload a single file as a raw request body
      description: The extension of `name` determines the stored extension.
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
          example: report.pdf
        - name: X-Expires-In
          in: header
          required: false
          description: "Delete the file after this long: 30m, 12h, 7d, 30d."
          schema: { type: string, example: 24h }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        "201":
          description: File stored
          headers:
            Location:
              description: URL of the first file stored by this request.
              schema: { type: string, format: uri }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "401": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
        "507": { $ref: "#/components/responses/Error" }
  /f/{id}/{name}:
    get:
      operationId: downloadFile
      summary: Download a file (public)
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, pattern: '^\d{8}-\d{6}-[0-9a-f]{32}$' }
        - name: name
          in: path
          required: true
          description: Cosmetic name; its extension must match the stored file.
          schema: { type: string }
        - name: dl
          in: query
          required: false
          description: "Any value forces `Content-Disposition: attachment`."
          schema: { type: string }
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "206": { description: Partial content (range request) }
        "304": { description: Not modified }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteFile
      summary: Delete a file
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /f/{stored}:
    get:
      operationId: downloadFileShort
      summary: Download using the short form `<id>.<ext>`
      security: []
      parameters:
        - name: stored
          in: path
          required: true
          schema: { type: string }
          example: 20260815-143022-8f4e2c91b7934b38a72d1c0e5b6a4f3d92.jpg
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteFileShort
      summary: Delete using the short form `<id>.<ext>`
      parameters:
        - name: stored
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /healthz:
    get:
      operationId: health
      summary: Liveness probe
      security: []
      responses:
        "200":
          description: Process is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: ok }
                  version: { type: string }
  /readyz:
    get:
      operationId: ready
      summary: Readiness probe (storage is writable)
      security: []
      responses:
        "200": { description: Ready }
        "503": { description: Storage is not writable }
  /stats:
    get:
      operationId: stats
      summary: Storage statistics
      responses:
        "200":
          description: Current usage
          content:
            application/json:
              schema:
                type: object
                properties:
                  files: { type: integer }
                  bytes: { type: integer }
                  bytes_human: { type: string }
                  quota_bytes: { type: integer, description: 0 means unlimited }
                  max_file: { type: integer }
                  uptime_s: { type: integer }
                  version: { type: string }
        "401": { $ref: "#/components/responses/Error" }
  /mcp:
    post:
      operationId: mcp
      summary: Model Context Protocol endpoint
      description: |
        JSON-RPC 2.0 in an HTTP POST, built on protocol revision `2026-07-28`.
        That revision has no handshake and no session: every request carries
        its own protocol version and client capabilities in `params._meta`, and
        the headers below mirror the body so an intermediary can route without
        parsing it. A body with no `id` is a notification and is answered with
        `202`.

        Methods: `server/discover`, `tools/list`, `tools/call`. Tools:
        `upload_file`, `delete_file`, `storage_stats`. The bytes of an upload
        arrive base64 encoded, so `upload_file` is capped well below
        `POST /upload`, which streams.

        A client built on an older revision (`2025-11-25` and earlier) opens
        with `initialize` instead and is answered in that shape, with `ping`
        and the same tools. The headers below are then not required, and no
        session id is assigned in either era.
      parameters:
        - name: MCP-Protocol-Version
          in: header
          required: false
          description: >
            Required on a `2026-07-28` request, where it must equal
            `params._meta["io.modelcontextprotocol/protocolVersion"]`. An older
            client names its version here after the handshake, or not at all.
          schema: { type: string, example: "2026-07-28" }
        - name: Mcp-Method
          in: header
          required: false
          description: >
            Required on a `2026-07-28` request, where it must equal the body's
            `method`.
          schema: { type: string, example: tools/call }
        - name: Mcp-Name
          in: header
          required: false
          description: >
            Required on a `2026-07-28` `tools/call`, where it must equal
            `params.name`. A value that is not plain ASCII is sent as
            `=?base64?<encoded>?=`.
          schema: { type: string, example: upload_file }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JSONRPCRequest" }
      responses:
        "200":
          description: The JSON-RPC response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JSONRPCResponse" }
        "202": { description: A notification was accepted }
        "400":
          description: >
            Malformed message (-32700, -32600), missing per-request metadata
            (-32602), a header that disagrees with the body (-32020), or an
            unsupported protocol version (-32022).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JSONRPCResponse" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { description: The Origin header is not allowed to reach this endpoint }
        "404":
          description: The method is not implemented (-32601)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JSONRPCResponse" }
        "405": { description: This revision has no GET stream and no DELETE }
        "413": { description: The message is larger than this endpoint holds }
        "429": { $ref: "#/components/responses/Error" }
  /llms.txt:
    get:
      operationId: llmsTxt
      summary: Plain-text API description for AI agents
      security: []
      responses:
        "200":
          description: Description of this instance
          content:
            text/plain:
              schema: { type: string }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Send `Authorization: Bearer <token>`. `X-API-Key: <token>` is accepted
        as an alternative.
  schemas:
    FileInfo:
      type: object
      required: [url, size_bytes]
      properties:
        url: { type: string, format: uri }
        name:
          type: string
          description: >
            The file name as sent, which the URL cannot give back: its
            cosmetic part is a slug.
        size_bytes: { type: integer, description: Bytes stored. }
        expires_at:
          type: string
          format: date-time
          description: >
            When the file deletes itself, present only when the upload asked
            for an expiry with X-Expires-In or ?expires=.
    UploadResponse:
      type: object
      required: [files]
      properties:
        files:
          type: array
          items: { $ref: "#/components/schemas/FileInfo" }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    JSONRPCRequest:
      type: object
      required: [jsonrpc, method]
      properties:
        jsonrpc: { type: string, const: "2.0" }
        id:
          description: Absent on a notification, and never null.
          oneOf: [{ type: string }, { type: integer }]
        method: { type: string, example: tools/call }
        params:
          type: object
          properties:
            name: { type: string, description: The tool, on tools/call. }
            arguments: { type: object }
            _meta:
              type: object
              description: >
                Required on a `2026-07-28` request, and absent from every
                older one, which is what tells the two apart.
              required:
                - io.modelcontextprotocol/protocolVersion
                - io.modelcontextprotocol/clientCapabilities
              properties:
                io.modelcontextprotocol/protocolVersion: { type: string, const: "2026-07-28" }
                io.modelcontextprotocol/clientCapabilities: { type: object }
                io.modelcontextprotocol/clientInfo:
                  type: object
                  properties:
                    name: { type: string }
                    version: { type: string }
    JSONRPCResponse:
      type: object
      required: [jsonrpc]
      properties:
        jsonrpc: { type: string, const: "2.0" }
        id:
          oneOf: [{ type: string }, { type: integer }]
        result:
          type: object
          required: [resultType]
          description: >
            Always carries `resultType`, and the server's identity in
            `_meta["io.modelcontextprotocol/serverInfo"]`. List results carry
            the caching hints `ttlMs` and `cacheScope` as well.
          properties:
            resultType: { type: string, const: complete }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: integer }
            message: { type: string }
            data: {}
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }